Allow JSON structured queries to be POSTed
This is a change needed by the analytics-dashboard, which currently
overwrites this file, as shown in this commit:
https://github.com/jianmin/analytics-dashboard/commit/3b1d58da2a7657b5187b3cd7b4ab26bfb4c49d2a

I would like to push this back into this group-by library, but I
discovered that it was not possible without converting the incoming JSON
structured query to a cts:query.

This, in turn, requires a dangerous call into MarkLogic's internal method
`Modules/MarkLogic/rest-api/lib/search-util.xqy#sut:search-from-json()`
to create an XML structured query. (We could duplicate the code, but
there seems to be a lot of it.) Then there is a bit of hackery,
calling `search:resolve` to recover the `cts:query` version of the
structured query.

For what it is worth, I also created an RFE to make these conversions
easier: https://rfetrack.marklogic.com/detail/3720